fix: разнести js по разным файлам, подправить стили
