+++ src/utils/tagify_templates.js
@@ -1,0 +1,32 @@
+export function tagTemplate(tagData){
+    return `
+        <tag title="${tagData.value}"
+                contenteditable='false'
+                spellcheck='false'
+                tabIndex="-1"
+                class="tagify__tag ${tagData.class ? tagData.class : ""}"
+                ${this.getAttributes(tagData)}>
+            <x title='' class='tagify__tag__removeBtn' role='button' aria-label='remove tag'></x>
+            <div>
+                <span class='tagify__tag-text'>${tagData.value}</span>
+            </div>
+        </tag>
+    `
+}
+
+export function suggestionItemTemplate(tagData){
+    return `
+        <div ${this.getAttributes(tagData)}
+            class='tagify__dropdown__item ${tagData.class ? tagData.class : ""}'
+            tabindex="0"
+            role="option">
+              <div class="main__checkbox">
+              <label class="main__checkbox-content">
+                <input type="checkbox" class="main__checkbox-input" />
+                <span class="main__checkbox-stile"></span>
+                ${tagData.value}
+              </label>
+              </div>
+        </div>
+    `
+}
